requirements.txt
coveralls==1.5.1
Cython
flake8==3.6.0
psutil==5.4.8
psycopg2-binary==2.7.6.1
pytest==4.0.0
pytest-cov==2.6.0
pytest-docker-py==1.1.0
pytest-timeout==1.3.3
python-qpid-proton==0.26.0
